[FIX] fix image export, use external product id as identifier, only export the image if the product was exported

+++ magentoerpconnect/product_images.py
@@ -25,9 +25,13 @@ import magerp_osv
 import mimetypes
 import netsvc
 from tools.translate import _
+import base64
 
 #TODO the option small_image, thumbnail, exclude, base_image, should be store diferently indeed this is not compatible with mutli instance (maybe serialized will be a good solution)
 #Moreover when a small is selected the flag on other image should be remove as magento does
+
+#TODO refactor all of this code and use the generic function from base_external_referentials.
+#To must useless line of code
 
 class product_images(magerp_osv.magerp_osv):
     _inherit = "product.images"
@@ -83,9 +87,9 @@ class product_images(magerp_osv.magerp_osv):
             return types
 
         #TODO update the image file
-        def update_image(image_name, image):
+        def update_image(product_extid, image_name, image):
             result = conn.call('catalog_product_attribute_media.update',
-                               [image.product_id.default_code,
+                               [product_extid,
                                 image_name,
                                 {'label':image.name,
                                  'exclude':image.exclude,
@@ -118,40 +122,49 @@ class product_images(magerp_osv.magerp_osv):
         while ids:
             product_images = self.browse_w_order(cr, uid, ids[:1000], context=context)
             for each in product_images:
-                need_to_be_created = True
-                ext_file_name = each.get_extid(context['external_referential_id'])
-                if ext_file_name: #If update
-                    try:
-                        logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "Updating %s's image: %s" %(each.product_id.default_code, each.name))
-                        result = update_image(ext_file_name, each)
-                        logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "%s's image updated with sucess: %s" %(each.product_id.default_code, each.name))
-                        need_to_be_created = False
-                    except Exception, e:
-                        logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("Error in connecting:%s") % (e))
-                        if not "Fault 103" in str(e):
-                            logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("Unknow error stop export"))
-                            raise
-                        else:
-                            #If the image was deleded in magento, the external name is automatically deleded before trying to re-create the image in magento
-                            model_data_ids = ir_model_data_obj.search(cr, uid, [('model', '=', self._name), ('res_id', '=', each.id), ('external_referential_id', '=', context['external_referential_id'])])
-                            if model_data_ids and len(model_data_ids) > 0:
-                                ir_model_data_obj.unlink(cr, uid, model_data_ids, context=context)
-                            logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("The image don't exist in magento, try to create it"))
-                if need_to_be_created:
-                    if each.product_id.default_code:
-                        logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "Sending %s's image: %s" %(each.product_id.default_code, each.name))
-                        result = conn.call('catalog_product_attribute_media.create',
-                                  [each.product_id.default_code,
-                                   {'file':{
-                                            'name':each.name,
-                                            'content': each.file,
-                                            'mime': each.link and each.url and mimetypes.guess_type(each.url)[0] or each.extention and mimetypes.guess_type(each.extention)[0] or 'image/jpeg',
-                                            }
-                                   }
-                                   ])
-                        self.create_external_id_vals(cr, uid, each.id, result, context['external_referential_id'], context=context)
-                        result = update_image(result, each)
-                        logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "%s's image send with sucess: %s" %(each.product_id.default_code, each.name))
+                product_extid = each.product_id.get_extid(context['external_referential_id'])
+                if not product_extid:
+                    logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "The product %s do not exist on magento" %(each.product_id.default_code))
+                else:
+                    need_to_be_created = True
+                    ext_file_name = each.get_extid(context['external_referential_id'])
+                    if ext_file_name: #If update
+                        try:
+                            logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "Updating %s's image: %s" %(each.product_id.default_code, each.name))
+                            result = update_image(ext_file_name, each)
+                            logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "%s's image updated with sucess: %s" %(each.product_id.default_code, each.name))
+                            need_to_be_created = False
+                        except Exception, e:
+                            logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("Error in connecting:%s") % (e))
+                            if not "Fault 103" in str(e):
+                                logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("Unknow error stop export"))
+                                raise
+                            else:
+                                #If the image was deleded in magento, the external name is automatically deleded before trying to re-create the image in magento
+                                model_data_ids = ir_model_data_obj.search(cr, uid, [('model', '=', self._name), ('res_id', '=', each.id), ('external_referential_id', '=', context['external_referential_id'])])
+                                if model_data_ids and len(model_data_ids) > 0:
+                                    ir_model_data_obj.unlink(cr, uid, model_data_ids, context=context)
+                                logger.notifyChannel(_("Magento Connection"), netsvc.LOG_ERROR, _("The image don't exist in magento, try to create it"))
+                    if need_to_be_created:
+                        if each.product_id.default_code:
+                            pas_ok = True
+                            suceed = False
+                            logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "Sending %s's image: %s" %(each.product_id.default_code, each.name))
+                            data = {
+                                'file':{
+                                    'name':each.name,
+                                    'content': each.file,
+                                    'mime': each.link and each.url and mimetypes.guess_type(each.url)[0] \
+                                            or each.extention and mimetypes.guess_type(each.name + each.extention)[0] \
+                                            or 'image/jpeg',
+                                    }
+                            }
+                            result = conn.call('catalog_product_attribute_media.create', [product_extid, data, False, 'id'])
+
+                            self.create_external_id_vals(cr, uid, each.id, result, context['external_referential_id'], context=context)
+                            result = update_image(product_extid, result, each)
+                            logger.notifyChannel('ext synchro', netsvc.LOG_INFO, "%s's image send with sucess: %s" %(each.product_id.default_code, each.name))
+
                 if image_2_date[each.id] > context['last_images_export_date']: #indeed if a product was created a long time ago and checked as exportable recently, the write date of the image can be far away in the past
                     self.pool.get('sale.shop').write(cr,uid,context['shop_id'],{'last_images_export_date':image_2_date[each.id]})
                 cr.commit()
